chore(system): remove unused files and dependencies
This change removes several files and dependencies that are no longer in use across the system configuration. The motivation for this cleanup is to reduce clutter and improve maintainability by ensuring only active and relevant components remain in the repository.

Specifically, the following files have been removed:
- `scripts/darwin-migrate-sequoia.sh`: A script intended for MacOS system updates, which is no longer necessary.
- `secrets/mullvad.age`: A secret file for Mullvad VPN, which is no longer referenced in the configuration.
- `system/home/process-compose.nix`: A configuration file for process management, which is no longer used.

Additionally, several packages and imports have been removed from the system configuration files, including:
- `todoist` and `cmatrix` from `system/darwin/homebrew.nix`.
- `b3sum`, `eternal-terminal`, `figlet`, `gh`, `gh-dash`, `git-bug`, `glow`, `pricehist`, and `tmuxp` from `system/default.nix`.
- `process-compose.nix` import from `system/home/default.nix`.
- `mullvad` configuration from `system/nixos/networking.nix`.
- `mullvad.enable` option from `system/options.nix`.

These removals align with the current state of the system and ensure that the configuration remains concise and focused on active functionality.